MS-1356 Combine logout use cases to have a single way to clear out data
