+++ infra/sync/src/test/java/com/simprints/infra/sync/config/usecase/LogoutUseCaseTest.kt
@@ -1,8 +1,11 @@
 package com.simprints.infra.sync.config.usecase
 
 import com.simprints.infra.authlogic.AuthManager
+import com.simprints.infra.enrolment.records.repository.EnrolmentRecordRepository
+import com.simprints.infra.enrolment.records.repository.local.migration.RealmToRoomMigrationFlagsStore
 import com.simprints.infra.sync.ScheduleCommand
 import com.simprints.infra.sync.SyncOrchestrator
+import com.simprints.testtools.common.coroutines.TestCoroutineRule
 import io.mockk.MockKAnnotations
 import io.mockk.coVerify
 import io.mockk.every
@@ -11,14 +14,24 @@ import io.mockk.verify
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
 
 class LogoutUseCaseTest {
+    @get:Rule
+    val testCoroutineRule = TestCoroutineRule()
+
     @MockK
     private lateinit var syncOrchestrator: SyncOrchestrator
 
     @MockK
     private lateinit var authManager: AuthManager
+
+    @MockK
+    private lateinit var flagsStore: RealmToRoomMigrationFlagsStore
+
+    @MockK
+    private lateinit var enrolmentRecordRepository: EnrolmentRecordRepository
 
     private lateinit var useCase: LogoutUseCase
 
@@ -30,6 +43,9 @@ class LogoutUseCaseTest {
         useCase = LogoutUseCase(
             syncOrchestrator = syncOrchestrator,
             authManager = authManager,
+            flagsStore = flagsStore,
+            enrolmentRecordRepository = enrolmentRecordRepository,
+            ioDispatcher = testCoroutineRule.testCoroutineDispatcher,
         )
     }
 
@@ -42,5 +58,7 @@ class LogoutUseCaseTest {
             syncOrchestrator.deleteEventSyncInfo()
             authManager.signOut()
         }
+        verify { flagsStore.clearMigrationFlags() }
+        coVerify { enrolmentRecordRepository.closeOpenDbConnection() }
     }
 }
